+++ addOns/reports/CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 ### Added
-- Added a new section to Traditional JSON Report with Requests and Responses for script diagnostics.
+- Added new sections to Traditional JSON Report with Requests and Responses for script diagnostics information.
 - Templates that include Insights now also identify the Insight that stopped the scan, if applicable.
 
 ### Fixed